+++ AnswerKing-NUnit-Ruaridh/Answer.King.IntegrationTests.csproj
@@ -29,6 +29,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Models\" />
     <Folder Include="Tests\__snapshots__\" />
   </ItemGroup>
