+++ src/Answer.King.Api/Answer.King.Api.csproj
@@ -59,8 +59,4 @@
     <None Include="appsettings.Production.json" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="db\" />
-  </ItemGroup>
-
 </Project>
